+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change log
 
+## [0.1.1] - 2024-10-02
+
+### Added
+
+- Documented some of the supported API calls.
+- Added original as output format
+- More options added
+
+### Changed
+
+- Some option names for api, for a list for options, run `matomo api --help`
+
+
 ## [0.1.0] - 2024-10-02
 
 ### Added